Lots of changes:
- Added a reference implementation of spmv and symgs.
- Tried to clean up the rest of the global/local int errors.
- Lots of small changes.

M    src/WriteProblem.cpp
M    src/GenerateProblem.cpp
M    src/main.cpp
M    src/dot.cpp
M    src/ExchangeHalo.cpp
M    src/CG.cpp
M    src/dot.hpp
M    src/Makefile
M    src/ReportResults.cpp
M    src/CG.hpp
M    src/Geometry.hpp
M    src/symgs.cpp
M    src/ComputeResidual.cpp
M    src/symgs.hpp
M    src/ComputeResidual.hpp
M    src/waxpby.cpp
M    src/waxpby.hpp
M    src/OptimizeMatrix.cpp
M    src/spmv.cpp
M    src/SparseMatrix.hpp
M    testing/CG/CGTest.cpp
M    testing/CG/Makefile


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@30 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -12,7 +12,7 @@
 #define GEOMETRY_HPP
 
 typedef int local_int_t;
-typedef int global_int_t;
+typedef long long global_int_t;
 
 // This is a data structure to contain all processor navigation information
